requirements.txt
flask
flask_bootstrap
flask_moment
flask-wtf
flask-sqlalchemy
flask-migrate
flask-mail
python-dotenv